First iteration of test cases

+++ src/main/java/com/employee/registry/employeeregistrydemo/EmployeeController.java
@@ -1,13 +1,11 @@
 package com.employee.registry.employeeregistrydemo;
 
-import java.net.URI;
 import java.util.List;
 import java.util.Optional;
 
 import javax.validation.Valid;
 
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -15,7 +13,7 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
-import org.springframework.web.servlet.support.ServletUriComponentsBuilder;
+
 
 @RestController
 public class EmployeeController {
@@ -53,15 +51,13 @@ public class EmployeeController {
 			return employee;
 		}
 	}
-	
-	// Add new employee
+
 	@PostMapping("/employees")
-	public ResponseEntity<Object> addEmployee(@Valid @RequestBody Employee employee) {
-		Employee savedEmployee = employeeRepository.save(employee);
-		URI location = ServletUriComponentsBuilder.fromCurrentRequest()
-				.path("/{id}").buildAndExpand(savedEmployee.getEmployeeId()).toUri();
-		return ResponseEntity.created(location).build();
-	}
+	public Employee addEmployee(@Valid @RequestBody Employee employee) {
+		 employeeRepository.save(employee);
+		return employee;
+	} 
+	
 	
 	// Update existing employee
 	@PutMapping("/employees/update/{id}")
@@ -80,7 +76,13 @@ public class EmployeeController {
 	// Delete employee by id
 	@DeleteMapping("/employees/id/{employeeId}")
 	public void deleteEmployeeById(@PathVariable int employeeId) {
-		employeeRepository.deleteById(employeeId);
+		if (employeeRepository.existsById(employeeId)) {
+			employeeRepository.deleteById(employeeId);
+		} else {
+			throw new UserNotFoundException("employee id: " +employeeId+" not found");
+		}
+		
+		
 	}
